+++ modules/validators/src/cpp/ValidatorsInternal.cpp
@@ -41,6 +41,7 @@
 #include "gtBuiltin.hpp"
 #include "geBuiltin.hpp"
 #include "ltBuiltin.hpp"
+#include "leBuiltin.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
@@ -298,6 +299,43 @@ mustBePositive(const ArrayOf& arg, int argPosition, bool asCaller)
     if (!isPositive) {
         std::wstring msg = invalidPositionMessage(argPosition) + _W("Value must be positive.");
         std::wstring id = _W("Nelson:validators:mustBePositive");
+        Error(msg, id, asCaller);
+    }
+}
+//=============================================================================
+void
+mustBeNonpositive(const ArrayOf& arg, int argPosition, bool asCaller)
+{
+    ArrayOfVector argIn(arg);
+    ArrayOfVector argOut = TypeGateway::isnumericBuiltin(_eval, 1, argIn);
+    bool isLogical = (arg.isLogical() || ClassName(arg) == "logical");
+    bool isNumeric = argOut[0].getContentAsLogicalScalar();
+    if (!isNumeric && !isLogical) {
+        std::wstring msg
+            = invalidPositionMessage(argPosition) + _W("Value must be numeric or logical.");
+        std::wstring id = _W("Nelson:validators:mustBeNumericOrLogical");
+        Error(msg, id, asCaller);
+    }
+
+    argOut = TypeGateway::isrealBuiltin(_eval, 1, argIn);
+    bool isReal = argOut[0].getContentAsLogicalScalar();
+    if (!isReal) {
+        std::wstring msg = invalidPositionMessage(argPosition) + _W("Value must be real.");
+        std::wstring id = _W("Nelson:validators:mustBeReal");
+        Error(msg, id, asCaller);
+    }
+    Dimensions dimsA = argIn[0].getDimensions();
+    Dimensions dimsV(1, dimsA.getElementCount());
+    ArrayOf asVector = argIn[0];
+    asVector.reshape(dimsV);
+    ArrayOfVector vAsArrayOfVector(asVector);
+    vAsArrayOfVector.push_back(ArrayOf::doubleConstructor(0));
+    argOut = ElementaryFunctionsGateway::leBuiltin(_eval, 1, vAsArrayOfVector);
+    argOut = ElementaryFunctionsGateway::allBuiltin(_eval, 1, argOut);
+    bool isNonpositive = argOut[0].getContentAsLogicalScalar();
+    if (!isNonpositive) {
+        std::wstring msg = invalidPositionMessage(argPosition) + _W("Value must be non positive.");
+        std::wstring id = _W("Nelson:validators:mustBeNonpositive");
         Error(msg, id, asCaller);
     }
 }