+++ modules/validators/src/include/Validators.hpp
@@ -73,11 +73,13 @@ NLSVALIDATORS_IMPEXP void
 mustBePositive(const ArrayOfVector& args, int argPosition);
 //=============================================================================
 NLSVALIDATORS_IMPEXP void
+mustBeNonpositive(const ArrayOfVector& args, int argPosition);
+//=============================================================================
+NLSVALIDATORS_IMPEXP void
 mustBeNonnegative(const ArrayOfVector& args, int argPosition);
 //=============================================================================
 NLSVALIDATORS_IMPEXP void
 mustBeNegative(const ArrayOfVector& args, int argPosition);
 //=============================================================================
-
 }
 //=============================================================================